Refactor the packing kernels.
