+++ pyfr/backends/base/__init__.py
@@ -1,9 +1,9 @@
 from pyfr.backends.base.backend import BaseBackend, MemoryInfo
-from pyfr.backends.base.kernels import (BaseKernelProvider,
-                                        BaseOrderedMetaKernel,
-                                        BasePointwiseKernelProvider,
-                                        BaseUnorderedMetaKernel, Kernel,
-                                        NotSuitableError, NullKernel)
+from pyfr.backends.base.provider import (BaseKernelProvider,
+                                         BaseOrderedMetaKernel,
+                                         BasePointwiseKernelProvider,
+                                         BaseUnorderedMetaKernel, Kernel,
+                                         NotSuitableError, NullKernel)
 from pyfr.backends.base.types import (ConstMatrix, Extent, Graph, Matrix,
                                       MatrixBase, MatrixSlice, StorageRegion,
                                       View, XchgMatrix, XchgView)
